Added research interfaces and classes, introduced player's balance and fixed GameCache

+++ Assets/Scripts/Economy/ICommodity.cs
@@ -4,6 +4,6 @@
     {
         string Name { get; }
 
-        decimal Price { get; set; }
+        int Price { get; set; }
     }
 }
